+++ GUI/ViewModel/LineItem.py
@@ -1,5 +1,7 @@
 from GUI.GuiHelpers import GetLineHeight
-from PySubtitle.Helpers import Linearise, UpdateFields
+from PySubtitle.Helpers import UpdateFields
+from PySubtitle.Helpers.Text import Linearise
+
 from GUI.ViewModel.ViewModelError import ViewModelError
 
 from PySide6.QtCore import Qt
