+++ PySubtitle/Helpers/__init__.py
@@ -1,37 +1,38 @@
 import datetime
 import os
 import logging
-import re
 import regex
-import unicodedata
 import srt
+
 from typing import List
 from PySubtitle.SubtitleError import SubtitleError
 
 def GetEnvBool(key, default=False):
+    """
+    Get a boolean value from an environment variable
+    """
     var = os.getenv(key, default)
     if var is not None:
         return str(var).lower() in ('true', 'yes', '1')
     return default
 
 def GetEnvFloat(name, default=None):
+    """
+    Get a float value from an environment variable
+    """
     value = os.getenv(name, default)
     if value is not None:
         return float(value)
     return default
 
 def GetEnvInteger(name, default=None):
+    """
+    Get an integer value from an environment variable
+    """
     value = os.getenv(name, default)
     if value is not None:
         return int(value)
     return default
-
-def Linearise(lines):
-    if not isinstance(lines, list):
-        lines = str(lines).split("\n")
-
-    lines = [ str(line).strip() for line in lines ]
-    return " | ".join(lines)
 
 def UpdateFields(item : dict, update: dict, fields : list[str]):
     """
@@ -61,38 +62,6 @@ def CreateSrtSubtitle(item) -> srt.Subtitle:
 
     return item
 
-def GetTimeDelta(time) -> datetime.timedelta:
-    if time is None:
-        return None
-
-    if isinstance(time, datetime.timedelta):
-        return time
-
-    try:
-        return srt.srt_timestamp_to_timedelta(str(time))
-
-    except Exception as e:
-        time = str(time).strip()
-        parts = regex.split('[:,]', time)
-
-        if len(parts) == 3:
-            if len(parts[-1]) == 3:
-                logging.warning(f"Adding hour to time '{time}'")
-                return GetTimeDelta(f"00:{parts[0]}:{parts[1]},{parts[2]}")
-            else:
-                logging.warning(f"Adding milliseconds to time '{time}'")
-                return GetTimeDelta(f"{parts[0]}:{parts[1]}:{parts[2],000}")
-
-        if len(parts) >= 4:
-            if len(parts[-1]) == 3:
-                logging.warning(f"Using last four parts of '{time}' as time")
-                return GetTimeDelta(f"{parts[-4]}:{parts[-3]}:{parts[-2]},{parts[-1]}")
-
-            logging.warning(f"Using first four parts of '{time}' as time")
-            return GetTimeDelta(f"{parts[0]}:{parts[1]}:{parts[2]},{parts[3]}")
-
-    raise ValueError(f"Unable to interpret time '{str(time)}'")
-
 def GetInputPath(filepath):
     if not filepath:
         return None
@@ -120,59 +89,6 @@ def GetOutputPath(filepath, language="translated"):
         basename = basename + language_suffix
 
     return os.path.join(os.path.dirname(filepath), f"{basename}.srt")
-
-def ParseTranslation(text):
-    """
-    Unpack a response from Chat GPT
-    """
-    text, summary = ExtractTag("summary", text)
-    text, synopsis = ExtractTag("synopsis", text)
-    text, scene = ExtractTag("scene", text)
-    text, names = ExtractTagList("names", text)
-
-    context = {
-        'summary': summary,
-        'scene': scene,
-        'synopsis': synopsis,
-        'names': names
-    }
-    return text, context
-
-def ExtractTag(tagname : str, text : str):
-    """
-    Look for an xml-like tag in the input text, and extract the contents.
-    """
-    open_tag = f"<{tagname}>"
-    close_tag = f"</{tagname}>"
-    empty_tag = f"<{tagname}/>"
-
-    text = text.replace(empty_tag, '')
-
-    end_index = text.rfind(close_tag)
-
-    if end_index == -1:
-        return text.strip(), None
-
-    start_index = text.rfind(open_tag, 0, end_index)
-    if start_index == -1:
-        #raise ValueError(f"Malformed {tagname} tags in response")
-        logging.warning(f"Malformed {tagname} tags in response")
-        return text.strip(), None
-
-    tag = text[start_index + len(open_tag):end_index].strip()
-    text_before = text[:start_index].strip()
-    text_after = text[end_index + len(close_tag):].strip()
-    text = '\n'.join([text_before, text_after]).strip()
-
-    return text, tag
-
-def ExtractTagList(tagname, text):
-    """
-    Look for an xml-like tag in the input text, and extract the contents as a comma or newline separated list.
-    """
-    text, tag = ExtractTag(tagname, text)
-    tag_list = [ item.strip() for item in regex.split("[\n,]", tag) ] if tag else []
-    return text, tag_list
 
 def MergeTranslations(lines, translated):
     """
@@ -228,34 +144,46 @@ def ResyncTranslatedLines(original_lines, translated_lines):
     elif num_original > num_translated:
         logging.warning(f"Number of lines in original and translated subtitles don't match. Synced {min_lines} lines.")
 
-
-def RemoveEmptyLines(lines):
-    return [line for line in lines if line.text and line.text.strip()]
-
-def RemoveWhitespaceAndPunctuation(string):
-    # Matches any punctuation, separator, or other Unicode character
-    pattern = r'[\p{P}\p{Z}\p{C}]'
-    stripped = regex.sub(pattern, '', string)
-
-    # Normalize Unicode characters to their canonical forms
-    normalized = unicodedata.normalize('NFC', stripped)
-    return normalized
-
-def ConvertWhitespaceBlocksToNewlines(text):
+def TimeDeltaToText(time: datetime.timedelta) -> str:
     """
-    Convert chinese commas or blocks of 3 or more spaces to newlines, unless there are newlines already
+    Convert a timedelta to a string
     """
-    if text and '\n' not in text:
-        text = re.sub(r' {3,}|\，\s*', '\n', text)
-    return text
+    return srt.timedelta_to_srt_timestamp(time).replace('00:', '') if time is not None else None
 
-def IsTextContentEqual(string1 : str, string2 : str):
-    if string1 and string2:
-        stripped1 = RemoveWhitespaceAndPunctuation(string1)
-        stripped2 = RemoveWhitespaceAndPunctuation(string2)
-        return stripped1 == stripped2
+def GetTimeDelta(time : datetime.timedelta | str | None) -> datetime.timedelta:
+    """
+    Ensure the input value is a timedelta, as best we can
+    """
+    if time is None:
+        return None
 
-    return string1 == string2
+    if isinstance(time, datetime.timedelta):
+        return time
+
+    try:
+        return srt.srt_timestamp_to_timedelta(str(time))
+
+    except Exception as e:
+        time = str(time).strip()
+        parts = regex.split('[:,]', time)
+
+        if len(parts) == 3:
+            if len(parts[-1]) == 3:
+                logging.warning(f"Adding hour to time '{time}'")
+                return GetTimeDelta(f"00:{parts[0]}:{parts[1]},{parts[2]}")
+            else:
+                logging.warning(f"Adding milliseconds to time '{time}'")
+                return GetTimeDelta(f"{parts[0]}:{parts[1]}:{parts[2],000}")
+
+        if len(parts) >= 4:
+            if len(parts[-1]) == 3:
+                logging.warning(f"Using last four parts of '{time}' as time")
+                return GetTimeDelta(f"{parts[-4]}:{parts[-3]}:{parts[-2]},{parts[-1]}")
+
+            logging.warning(f"Using first four parts of '{time}' as time")
+            return GetTimeDelta(f"{parts[0]}:{parts[1]}:{parts[2]},{parts[3]}")
+
+    raise ValueError(f"Unable to interpret time '{str(time)}'")
 
 def ParseDelayFromHeader(value : str):
     """
@@ -297,54 +225,6 @@ def FormatMessages(messages):
         lines.append("")
 
     return '\n'.join(lines)
-
-def LimitTextLength(text, max_length):
-    text = text.strip()
-
-    if len(text) <= max_length:
-        return text
-
-    pattern = r'\.|\?|!'
-    matches = [(m.start(), m.group()) for m in re.finditer(pattern, text)]
-
-    # Find the closest match to the max length, if any
-    for position, match in reversed(matches):
-        if position <= max_length:
-            return text[:position + 1]
-
-    # If no sentence end is found within the limit, cut at the nearest whitespace
-    nearest_space = text.rfind(' ', 0, max_length)
-    if nearest_space != -1:
-        return text[:nearest_space] + '...'
-    else:
-        # As a last resort, cut directly at the max length
-        return text[:max_length] + '...'
-
-def SanitiseSummary(summary : str, movie_name : str = None, max_summary_length : int = None):
-    """
-    Remove trivial parts of summary text
-    """
-    if not summary:
-        return None
-
-    summary = re.sub(r'^(?:(?:Scene|Batch)[\s\d:\-]*)+', '', summary, flags=re.IGNORECASE)
-    summary = summary.replace("Summary of the batch", "")
-    summary = summary.replace("Summary of the scene", "")
-
-    if movie_name:
-        # Remove movie name and any connectors (-,: or whitespace)
-        summary = re.sub(r'^' + re.escape(movie_name) + r'\s*[:\-]\s*', '', summary)
-
-    summary = summary.strip()
-    original_len = len(summary)
-
-    if max_summary_length:
-        summary = LimitTextLength(summary, max_summary_length)
-
-    if len(summary) != original_len:
-        logging.info(f"Summary was truncated from {original_len} to {len(summary)} characters")
-
-    return summary or None
 
 def FormatErrorMessages(errors : List[SubtitleError]):
     """
